Make it easy to set a custom value generator
Issue #5537. Also see #5303.

This change added fluent API set set a custom value generator for use on a property. The value generator will be used regardless of provider and does not do anything special to ensure Migrations or the update pipeline are in sync with what is happening, but is useful to make a change to a value generator when these things do not matter or are already setup correctly.